requirements.txt
connexion>=0.8.6
kafka-python
stups-tokens>=0.3
prometheus
kazoo
gunicorn
eventlet
setproctitle
cachetools